Add files via upload
SQL Assignment 1

What is a relational database management system (RDBMS)? What are the advantages of a database management system over a file system?

In a database management system, explain the ACID properties.

Explain the concept of normalization.

Explain the many types of query languages used in relational databases. DQL, DML, DCL, and DDL are some examples.

What is the difference between the main key and a composite key? Give instances of how primary key and composite are used.

Create a table with a primary key, a column default value, and a column unique constraint in SQL.